Rename `ember-cli` to `ember_cli` in Ruby
Closes [#282].

[#282]: https://github.com/thoughtbot/ember-cli-rails/issues/282
