+++ app/controller/ember_cli/ember_controller.rb
