+++ lib/ember_cli/capture.rb
