+++ lib/ember_cli/command.rb
