+++ lib/ember_cli/configuration.rb
