+++ lib/ember_cli/constraint.rb
