+++ lib/ember_cli/controller_extension.rb
