+++ lib/ember_cli/errors.rb
