+++ lib/ember_cli/helpers.rb
