+++ lib/ember_cli/html_page.rb
