+++ lib/ember_cli/path_set.rb
