+++ lib/ember_cli/version.rb
