+++ spec/lib/ember_cli/app_spec.rb
