+++ spec/lib/ember_cli/html_page_spec.rb
